app/agents/boleto/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
azure-storage-blob==12.21.0
python-multipart==0.0.9
